DTO, services, prometheus, controllers

+++ src/main/java/com/tinderforprojects/tinder/model/photo/PhotoService.java
@@ -2,7 +2,7 @@ package com.tinderforprojects.tinder.model.photo;
 
 public interface PhotoService {
 
-    void upload(byte[] image, Integer id, String type);
+    void upload(byte[] image, Long id, String type);
 
     byte[] download(String hash);
